# docs/phase-1/mango-mdu-openapi.yaml
openapi: 3.0.3
info:
  title: Mango MDU Service API
  version: 1.0.0
  description: |
    Northbound API exposed by MDU for Phase 1 workflows.

    This contract defines the API consumed by northbound clients such as the MDU UI.
    It intentionally normalizes downstream platform behavior into stable MDU-owned resources.

    Phase 1 integrations:
    - OWSEC: authentication, token validation, and user identity/user CRUD handled directly outside MDU
    - PROV: entities, venues, effective scope resolution

    Phase 1 scope:
    - session/access context, hierarchy/entity/venue wrappers, and user-access orchestration only

    Explicit non-goal:
    - billing features are out-of-scope for Phase 1
servers:
  - url: /api/v1
    description: Relative API base path
...
security:
  - bearerAuth: []
paths:
...
  /entities/{entityId}/venues:
    get:
      tags: [Venues]
      summary: List venues under an entity
      operationId: listEntityVenues
      parameters:
        - $ref: '#/components/parameters/XRequestIdHeader'
        - $ref: '#/components/parameters/XCorrelationIdHeader'
        - $ref: '#/components/parameters/EntityIdPath'
        - $ref: '#/components/parameters/LimitQuery'
        - $ref: '#/components/parameters/OffsetQuery'
      responses:
        '200':
          description: Venue list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VenueListResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '503':
          $ref: '#/components/responses/BackendUnavailableError'
    post:
      tags: [Venues]
      summary: Create venue under an entity
      operationId: createVenue
      parameters:
        - $ref: '#/components/parameters/XRequestIdHeader'
        - $ref: '#/components/parameters/XCorrelationIdHeader'
        - $ref: '#/components/parameters/EntityIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVenueRequest'
      responses:
        '201':
          description: Venue created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VenueDetail'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '503':
          $ref: '#/components/responses/BackendUnavailableError'
  /venues/{venueId}:
    get:
      tags: [Venues]
      summary: Get venue detail
      operationId: getVenue
      parameters:
        - $ref: '#/components/parameters/XRequestIdHeader'
        - $ref: '#/components/parameters/XCorrelationIdHeader'
        - $ref: '#/components/parameters/VenueIdPath'
      responses:
        '200':
          description: Venue detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VenueDetail'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '503':
          $ref: '#/components/responses/BackendUnavailableError'
    put:
      tags: [Venues]
      summary: Update a venue
      operationId: updateVenue
      parameters:
        - $ref: '#/components/parameters/XRequestIdHeader'
        - $ref: '#/components/parameters/XCorrelationIdHeader'
        - $ref: '#/components/parameters/VenueIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVenueRequest'
      responses:
        '200':
          description: Venue updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VenueDetail'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '503':
          $ref: '#/components/responses/BackendUnavailableError'
    delete:
      tags: [Venues]
      summary: Delete a venue
      operationId: deleteVenue
      parameters:
        - $ref: '#/components/parameters/XRequestIdHeader'
        - $ref: '#/components/parameters/XCorrelationIdHeader'
        - $ref: '#/components/parameters/VenueIdPath'
      responses:
        '204':
          description: Venue deleted
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '503':
          $ref: '#/components/responses/BackendUnavailableError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OWSEC-issued user bearer token authentication. Note that token validation is conditionally enforced on the public interface and can be bypassed for local development and testing by setting the environment variable AUTH_ENABLED=false.
  parameters:
    XRequestIdHeader:
      name: X-Request-Id
      in: header
      required: false
      description: Unique tracking identifier for this specific HTTP request. If not supplied, the service will generate one.
      schema:
        type: string
    XCorrelationIdHeader:
      name: X-Correlation-Id
      in: header
      required: false
      description: Distributed transaction identifier linking requests across multiple microservices. If not supplied, the service will fall back to using X-Request-Id (or generate one) to preserve trace context.
      schema:
        type: string
    EntityIdPath:
      name: entityId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Id'
    VenueIdPath:
      name: venueId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Id'
...
    LimitQuery:
      name: limit
      in: query
      required: false
      description: Maximum number of items to return.
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    OffsetQuery:
      name: offset
      in: query
      required: false
      description: Number of items to skip.
      schema:
        type: integer
        default: 0
        minimum: 0
...
  responses:
    BadRequestError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            badRequestDefault:
              summary: Validation error default
              value:
                ErrorCode: 400
                ErrorDetails: "One or more request parameters are invalid."
                ErrorDescription: "Bad Request"
    UnauthorizedError:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            unauthorizedDefault:
              summary: Unauthorized access error
              value:
                ErrorCode: 401
                ErrorDetails: "Bearer token is missing, expired, or invalid."
                ErrorDescription: "Unauthorized"
    ForbiddenError:
      description: Caller is not allowed to perform this action in the resolved scope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            forbiddenDefault:
              summary: Forbidden access error
              value:
                ErrorCode: 403
                ErrorDetails: "You do not have permission to access resources in the resolved scope."
                ErrorDescription: "Forbidden"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            notFoundDefault:
              summary: Resource not found error
              value:
                ErrorCode: 404
                ErrorDetails: "The requested resource could not be found."
                ErrorDescription: "Not Found"
    ConflictError:
      description: Current state conflicts with the requested operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            conflictDefault:
              summary: Conflict error default
              value:
                ErrorCode: 409
                ErrorDetails: "The operation conflicts with the current state of the resource."
                ErrorDescription: "Conflict"
    BackendUnavailableError:
      description: One or more required downstream services are unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            backendUnavailableDefault:
              summary: Backend service unavailable
              value:
                ErrorCode: 503
                ErrorDetails: "A required downstream service (e.g. PROV or OWSEC) is currently unreachable or timed out."
                ErrorDescription: "Service Unavailable"
  schemas:
    Id:
      type: string
      description: UUID or service-owned opaque identifier
    IsoDateTime:
      type: string
      format: date-time
    ListMetadata:
      type: object
      required: [total, limit, offset]
      properties:
        total:
          type: integer
          description: Total count of items matching the filter.
        limit:
          type: integer
          description: Maximum number of items requested.
        offset:
          type: integer
          description: Offset of the returned items.
    HierarchyNodeType:
      type: string
      enum: [operator, entity, venue]
...
    ApiError:
      type: object
      required: [ErrorCode, ErrorDetails, ErrorDescription]
      properties:
        ErrorCode:
          type: integer
          description: Numeric error code (typically matching the HTTP status code or specific REST sub-code).
        ErrorDetails:
          type: string
          description: Detailed explanation or context of the error.
        ErrorDescription:
          type: string
          description: High-level classification of the error category.
      example:
        ErrorCode: 400
        ErrorDetails: "The requested field 'email' is invalid."
        ErrorDescription: "Bad Request"
    ScopePathItem:
      type: object
      required: [id, type, name]
      properties:
        id:
          $ref: '#/components/schemas/Id'
        type:
          $ref: '#/components/schemas/HierarchyNodeType'
        name:
          type: string
...
    VenueSummary:
      type: object
      required: [id, name, entityId, path]
      properties:
        id:
          $ref: '#/components/schemas/Id'
        name:
          type: string
        entityId:
          $ref: '#/components/schemas/Id'
        parentVenueId:
          allOf:
            - $ref: '#/components/schemas/Id'
          nullable: true
        path:
          type: array
          items:
            $ref: '#/components/schemas/ScopePathItem'
        deviceCount:
          type: integer
    VenueDetail:
      allOf:
        - $ref: '#/components/schemas/VenueSummary'
        - type: object
          properties:
            description:
              type: string
            managementPolicyId:
              $ref: '#/components/schemas/Id'
            managementRoleIds:
              type: array
              items:
                $ref: '#/components/schemas/Id'
            createdAt:
              $ref: '#/components/schemas/IsoDateTime'
            updatedAt:
              $ref: '#/components/schemas/IsoDateTime'
      example:
        id: "423e4567-e89b-12d3-a456-426614174000"
        name: "Grand MDU Complex"
        entityId: "223e4567-e89b-12d3-a456-426614174000"
        parentVenueId: null
        path:
          - id: "223e4567-e89b-12d3-a456-426614174000"
            name: "Regional Operations"
            type: "entity"
        deviceCount: 45
        description: "Main residential complex venue"
        managementPolicyId: "523e4567-e89b-12d3-a456-426614174000"
        managementRoleIds:
          - "623e4567-e89b-12d3-a456-426614174000"
        createdAt: "2026-06-30T12:00:00Z"
        updatedAt: "2026-06-30T12:00:00Z"
    VenueListResponse:
      type: object
      required: [items, metadata]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/VenueSummary'
        metadata:
          $ref: '#/components/schemas/ListMetadata'
    CreateVenueRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        parentVenueId:
          $ref: '#/components/schemas/Id'
    UpdateVenueRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
...